Minor Fixes

+++ app.js
@@ -75,7 +75,7 @@ app.get('/api/employeeDetails', (req, res) => {
             }
         });
     } else {
-        failureResponse(503,'Facing technical issues with Database, Please try again later.', res)
+        failureResponse(503, 'Facing technical issues with Database, Please try again later.', res)
     }
 });
 
@@ -93,7 +93,7 @@ app.get('/api/employeeDetails/:employeeID', (req, res) => {
             }
         });
     } else {
-        failureResponse(503,'Facing technical issues with Database, Please try again later.', res)
+        failureResponse(503, 'Facing technical issues with Database, Please try again later.', res)
     }
 });
 
@@ -102,30 +102,30 @@ app.post('/api/addEmployee', (req, res) => {
     if (isSQLConfigurationSuccess) {
         // Validate request data
         if (!req.body.employeeID || req.body.employeeID.length < 1) {
-            failureResponse(400,'Employee ID should not be empty.' , res)
+            failureResponse(400, 'Employee ID should not be empty.', res)
             return;
         } else if (!req.body.employeeName || req.body.employeeName.length < 3) {
-            failureResponse(400,'Please provide employee name with atleast 3 characters long.' , res)
+            failureResponse(400, 'Please provide employee name with atleast 3 characters long.', res)
             return;
         }
         // Check if employee with given id already exists
         const empData = EmployeeData.find(e => e.employeeID === parseInt(req.body.employeeID))
         if (empData) {
             // Employee with given ID already exists
-            failureResponse(400,'The employee with given employeeID already exists.' , res)
+            failureResponse(400, 'The employee with given employeeID already exists.', res)
             return;
         }
         // Input Data on Database
         connection.query("INSERT INTO `EmployeeManagement`.`EmployeeData` (`Employee ID`, `EmployeeName`) VALUES ('" + parseInt(req.body.employeeID) + "', '" + req.body.employeeName + "');", (err, result) => {
             if (err) {
-                failureResponse(400,'Failed to insert data into database, Please check your request and try again.' , res)
+                failureResponse(400, 'Failed to insert data into database, Please check your request and try again.', res)
             } else {
                 res.status(200).send('Employee Data Inserted Successfully !!');
             }
         });
         //res.send(employeeData);
     } else {
-        failureResponse(503,'Facing technical issues with Database, Please try again later.', res)
+        failureResponse(503, 'Facing technical issues with Database, Please try again later.', res)
     }
 });
 
@@ -135,28 +135,41 @@ app.put('/api/updateEmployeeData', (req, res) => {
 
         connection.query("UPDATE `EmployeeData` SET EmployeeName = '" + req.body.employeeName + "' WHERE `Employee ID` = " + req.body.employeeID, (err, result) => {
             if (err) {
-                failureResponse(400,'Failed to update data into database, Please check your request and try again.' , res)
+                failureResponse(400, 'Failed to update data into database, Please check your request and try again.', res)
             } else {
                 res.status(200).send('Employee Data Updated Successfully !!');
             }
         });
     } else {
-        failureResponse(503,'Facing technical issues with Database, Please try again later.', res)
+        failureResponse(503, 'Facing technical issues with Database, Please try again later.', res)
     }
 });
 
 // API to delete employee data
-app.delete('/api/deleteEmployeeData', (req, res) => {
+app.delete('/api/deleteEmployeeData', async (req, res) => {
     if (isSQLConfigurationSuccess) {
-        connection.query("DELETE FROM `EmployeeData` WHERE `Employee ID` = " + req.body.employeeID, (err, result) => {
+        const query = "SELECT EmployeeName FROM EmployeeData WHERE `Employee ID`= " + req.body.employeeID
+        connection.query(query, function (err, row) {
             if (err) {
-                failureResponse(400,'Failed to delete data into database, Please check your request and try again.' , res)
+                failureResponse(400, 'Employee with the given ID does not exist', res)
+                return;
             } else {
-                res.status(200).send('Employee Data Deleted Successfully !!');
+                if (row.length == 0 || row === undefined) {
+                    failureResponse(400, 'Employee with the given ID does not exist', res)
+                    return;
+                }
+                // Employee Exist Perform Delete operation
+                connection.query("DELETE FROM `EmployeeData` WHERE `Employee ID` = " + req.body.employeeID, (err, result) => {
+                    if (err) {
+                        failureResponse(400, 'Failed to delete data into database, Please check your request and try again.', res)
+                    } else {
+                        res.status(200).send('Employee Data Deleted Successfully !!');
+                    }
+                });
             }
         });
     } else {
-        failureResponse(503,'Facing technical issues with Database, Please try again later.', res)
+        failureResponse(503, 'Facing technical issues with Database, Please try again later.', res)
     }
 });
